Share common static forwader tests
